move page table initialization to the hardware dependent part
