+++ src/macos/x86_64/uhyve.rs
@@ -1,11 +1,14 @@
 use crate::arch::x86_64::BootInfo;
+use crate::consts::*;
 use crate::macos::x86_64::ioapic::IoApic;
 use crate::macos::x86_64::vcpu::*;
 use crate::vm::HypervisorResult;
 use crate::vm::{Parameter, Vm};
+use crate::x86_64::paging::*;
 use libc;
 use libc::c_void;
 use log::debug;
+use std::mem;
 use std::net::Ipv4Addr;
 use std::path::Path;
 use std::path::PathBuf;
@@ -13,6 +16,15 @@ use std::ptr;
 use std::ptr::read_volatile;
 use std::sync::{Arc, Mutex};
 use xhypervisor::{create_vm, map_mem, unmap_mem, MemPerm};
+
+// Constructor for a conventional segment GDT (or LDT) entry
+fn create_gdt_entry(flags: u64, base: u64, limit: u64) -> u64 {
+	((base & 0xff000000u64) << (56 - 24))
+		| ((flags & 0x0000f0ffu64) << 40)
+		| ((limit & 0x000f0000u64) << (48 - 16))
+		| ((base & 0x00ffffffu64) << 16)
+		| (limit & 0x0000ffffu64)
+}
 
 pub struct Uhyve {
 	offset: u64,
@@ -153,6 +165,57 @@ impl Vm for Uhyve {
 			0
 		} else {
 			unsafe { read_volatile(&(*self.boot_info).cpu_online) }
+		}
+	}
+
+	/// Initialize the page tables for the guest
+	fn init_guest_mem(&self) {
+		debug!("Initialize guest memory");
+
+		let (mem_addr, _) = self.guest_mem();
+
+		unsafe {
+			let pml4 = &mut *((mem_addr as u64 + BOOT_PML4) as *mut PageTable);
+			let pdpte = &mut *((mem_addr as u64 + BOOT_PDPTE) as *mut PageTable);
+			let pde = &mut *((mem_addr as u64 + BOOT_PDE) as *mut PageTable);
+			let gdt_entry: u64 = mem_addr as u64 + BOOT_GDT;
+
+			// initialize GDT
+			*((gdt_entry) as *mut u64) = create_gdt_entry(0, 0, 0);
+			*((gdt_entry + mem::size_of::<*mut u64>() as u64) as *mut u64) =
+				create_gdt_entry(0xA09B, 0, 0xFFFFF); /* code */
+			*((gdt_entry + 2 * mem::size_of::<*mut u64>() as u64) as *mut u64) =
+				create_gdt_entry(0xC093, 0, 0xFFFFF); /* data */
+
+			/* For simplicity we currently use 2MB pages and only a single
+			PML4/PDPTE/PDE. */
+
+			// per default is the memory zeroed, which we allocate by the system call mmap
+			/*libc::memset(pml4 as *mut _ as *mut libc::c_void, 0x00, PAGE_SIZE);
+			libc::memset(pdpte as *mut _ as *mut libc::c_void, 0x00, PAGE_SIZE);
+			libc::memset(pde as *mut _ as *mut libc::c_void, 0x00, PAGE_SIZE);*/
+
+			pml4.entries[0].set(
+				BOOT_PDPTE as usize,
+				PageTableEntryFlags::PRESENT | PageTableEntryFlags::WRITABLE,
+			);
+			pml4.entries[511].set(
+				BOOT_PML4 as usize,
+				PageTableEntryFlags::PRESENT | PageTableEntryFlags::WRITABLE,
+			);
+			pdpte.entries[0].set(
+				BOOT_PDE as usize,
+				PageTableEntryFlags::PRESENT | PageTableEntryFlags::WRITABLE,
+			);
+
+			for i in 0..512 {
+				pde.entries[i].set(
+					i * LargePageSize::SIZE,
+					PageTableEntryFlags::PRESENT
+						| PageTableEntryFlags::WRITABLE
+						| PageTableEntryFlags::HUGE_PAGE,
+				);
+			}
 		}
 	}
 }
